+++ src/liquid_parquet/src/reader/plantime/opener.rs
@@ -58,6 +58,7 @@ pub struct LiquidParquetOpener {
 }
 
 impl LiquidParquetOpener {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         partition_index: usize,
         projection: Arc<[usize]>,
